+++ tests/testthat/test-quick_delin.R
@@ -1,7 +1,0 @@
-test_that("quick_delin", {
-  skip_if_not(interactive())
-
-  stimuli <- demo_stim() |> remove_tem()
-  x <- quick_delin(stimuli, 2)
-  plot(x, pt.plot = T)
-})